requirements.txt
openai==0.27.0
SQLAlchemy==2.0.10
clickhouse-sqlalchemy==0.1.8
pandas==2.0.1
python-telegram-bot==20.3
python-dotenv==1.0.0
openpyxl==3.1.2
matplotlib==3.7.1
python-dotenv==1.0.1
